VirtNotifications: move ID calculation to the action class
Use inheritance to compute the action id to send via the websocket
rather than an ungly classname-based map.
